Fix use of LoggingRenderer after bad merge
